Move async into its own module
