+++ lib/CMakeLists.txt
@@ -123,7 +123,7 @@ target_link_libraries(arango_lightweight
          arango_inspection
          absl::flat_hash_set
          absl::synchronization
-         arango_coroutine_registry
+         arango_async
   PRIVATE date_interface
           fmt
           arango_assertions
@@ -298,6 +298,7 @@ if (USE_ENTERPRISE)
 endif()
 
 add_subdirectory(Actor)
+add_subdirectory(Async)
 add_subdirectory(Futures)
 add_subdirectory(Geo)
 if (USE_V8)
@@ -308,4 +309,3 @@ add_subdirectory(CrashHandler)
 add_subdirectory(Assertions)
 add_subdirectory(Inspection)
 add_subdirectory(BuildId)
-add_subdirectory(Basics)
