+++ tests/Async/CMakeLists.txt
@@ -1,0 +1,3 @@
+target_sources(arangodbtests
+  PRIVATE
+  AsyncTest.cpp)
