+++ tests/Basics/CMakeLists.txt
@@ -1,10 +1,8 @@
 target_sources(arangodbtests
   PRIVATE
-  AsyncTest.cpp
   ConversionsTest.cpp
   CsvTest.cpp
   datetime.cpp
-  ExpectedTests.cpp
   FilesTest.cpp
   FpConvTest.cpp
   FutureSharedLockTest.cpp
@@ -40,5 +38,3 @@ target_sources(arangodbtests
   UnshackledMutexTest.cpp
   VelocyPackHelperTest.cpp
   VelocyPackHelperCompareRandomTest.cpp)
-
-add_subdirectory(coroutine)
