--- tests/Basics/coroutine/CMakeLists.txt
@@ -1,4 +1,0 @@
-target_sources(arangodbtests
-  PRIVATE
-	ThreadRegistryTest.cpp
-	RegistryTest.cpp)
